refactor(agentic): mechanically split mod.rs into facade, builder, and helpers (wave12 step2)
